new-relic-deployment-marker/requirements.txt
cerberus
requests
bitbucket-pipes-toolkit
PyYAML